Renamed config.ini to config.ini.example.
As the config will be installation specific, we don't want to pass around a config that's "broken" for certain installations.  And we want to be able to git pull on the live server without then having to copy/edit files each time.

So we have our ini file as a .example file, then on each installation we copy that and create our own config.ini.  We then add config.ini to the git ignore list, so it doesn't get committed/pulled/pushed.