More work done on generating test data.
Now that the FetchTree function is complete I've been able to take the next step and get a list of all the 'leaf' categories for which we're going to generate data (parent categories have their values calculated).

We're also now defining and looping through the date range for which the test data will be generated.  Next step is to create the library functions required to create that test data.